HW6: Fixed BFS and output according to hw6extra1testOutput1
So it just so happened that the other order of doing BFS made the right output for the first 2 testOutputs but the real order we add to the array is rldu and since we are visiting the nodes after putting them in a queue, it does go udlr. There we're a few other relatively minor differences between the outputs that are now fixed.